test: working on test cases 3
