+++ pkg/verifier/notation/notationrevocationfactory.go
@@ -11,22 +11,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package revocation
+package notation
 
 import (
 	"net/http"
 
 	"github.com/notaryproject/notation-core-go/revocation"
 	corecrl "github.com/notaryproject/notation-core-go/revocation/crl"
+	"github.com/notaryproject/notation-go/verifier/crl"
 )
 
-type RevocationFactory interface {
-	// NewFetcher returns a new fetcher instance
-	NewFetcher(*http.Client) (*corecrl.HTTPFetcher, error)
+type Notationrevocationfactory struct{}
 
-	// NewFileCache returns a new file cache instance
-	NewFileCache(string) (corecrl.Cache, error)
+// NewFetcher returns a new fetcher instance
+func (f *Notationrevocationfactory) NewFetcher(client *http.Client) (*corecrl.HTTPFetcher, error) {
+	return corecrl.NewHTTPFetcher(client)
+}
 
-	// NewValidator returns a new validator instance
-	NewValidator(revocation.Options) (revocation.Validator, error)
+// NewFileCache returns a new file cache instance
+func (f *Notationrevocationfactory) NewFileCache(cacheDir string) (corecrl.Cache, error) {
+	return crl.NewFileCache(cacheDir)
+}
+
+// NewValidator returns a new validator instance
+func (f *Notationrevocationfactory) NewValidator(opts revocation.Options) (revocation.Validator, error) {
+	return revocation.NewWithOptions(opts)
 }
